Move exception tests from integration to unit tests
