+++ atlasdb-cassandra/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraClientPoolTest.java
@@ -18,7 +18,9 @@ package com.palantir.atlasdb.keyvalue.cassandra;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -31,6 +33,10 @@ import java.util.NoSuchElementException;
 import java.util.Optional;
 
 import org.apache.cassandra.thrift.Cassandra;
+import org.apache.cassandra.thrift.InvalidRequestException;
+import org.apache.cassandra.thrift.TimedOutException;
+import org.apache.cassandra.thrift.UnavailableException;
+import org.apache.thrift.transport.TTransportException;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
@@ -212,6 +218,39 @@ public class CassandraClientPoolTest {
         Mockito.verify(cassandraClientPool.getCurrentPools().get(host), Mockito.times(numAttempts))
                 .runWithPooledResource(
                         Mockito.<FunctionCheckedException<Cassandra.Client, Object, RuntimeException>>any());
+    }
+
+    @Test
+    public void testIsConnectionException() {
+        assertFalse(CassandraClientPoolImpl.isConnectionException(new TimedOutException()));
+        assertFalse(CassandraClientPoolImpl.isConnectionException(new TTransportException()));
+        assertTrue(CassandraClientPoolImpl.isConnectionException(new TTransportException(
+                new SocketTimeoutException())));
+    }
+
+    @Test
+    public void testIsRetriableException() {
+        assertTrue(CassandraClientPoolImpl.isRetriableException(new TimedOutException()));
+        assertTrue(CassandraClientPoolImpl.isRetriableException(new TTransportException()));
+        assertTrue(CassandraClientPoolImpl.isRetriableException(new TTransportException(new SocketTimeoutException())));
+    }
+
+    @Test
+    public void testIsRetriableWithBackoffException() {
+        assertTrue(CassandraClientPoolImpl.isRetriableWithBackoffException(new NoSuchElementException()));
+        assertTrue(CassandraClientPoolImpl.isRetriableWithBackoffException(new UnavailableException()));
+        assertTrue(CassandraClientPoolImpl.isRetriableWithBackoffException(
+                new TTransportException(new SocketTimeoutException())));
+        assertTrue(CassandraClientPoolImpl.isRetriableWithBackoffException(
+                new TTransportException(new UnavailableException())));
+    }
+
+    @Test
+    public void testIsFastFailoverException() {
+        assertFalse(CassandraClientPoolImpl.isRetriableWithBackoffException(new InvalidRequestException()));
+        assertFalse(CassandraClientPoolImpl.isRetriableException(new InvalidRequestException()));
+        assertFalse(CassandraClientPoolImpl.isConnectionException(new InvalidRequestException()));
+        assertTrue(CassandraClientPoolImpl.isFastFailoverException(new InvalidRequestException()));
     }
 
     private CassandraClientPoolImpl clientPoolWithServers(ImmutableSet<InetSocketAddress> servers) {
